bug with database update (bad request)w

+++ archaeologyProject.py
@@ -16,11 +16,12 @@ class Site(db.Model):
     name = db.Column(db.String(250))
     toponim = db.Column(db.String(250))
     type_of_site= db.Column(db.String(250), nullable=False)
-#    oblast = db.Column(db.String(250), nullable=False)
-#    rajon = db.Column(db.String(250), nullable=False)
+    oblast = db.Column(db.String(250), nullable=False)
+    rajon = db.Column(db.String(250), nullable=False)
 #    punkt = db.Column(db.String(250), nullable=False)
-#    prymitky = db.Column(db.String(3000))
-##    kultnal = db.Column(db.String(250), nullable=False)
+#    pryvjazka = db.Column(db.String(250), nullable=False)
+#    kultnal = db.Column(db.String(250), nullable=False)
+#    localgr = db.Column(db.String(250), nullable=False)
 #    chron = db.Column(db.String(250), nullable=False)
 #    nadijnist = db.Column(db.String(250), nullable=False)
 #    rozkop = db.Column(db.String(1000), nullable=False)
@@ -29,24 +30,25 @@ class Site(db.Model):
 #    kartograph = db.Column(db.String(20), nullable=False)
 #    coord = db.Column(db.String(50), nullable=False)
 #    tochkart = db.Column(db.String(20), nullable=False)
+#    basejn = db.Column(db.String(50), nullable=False)
 #    toppotype = db.Column(db.String(30), nullable=False)
 #    geomorform = db.Column(db.String(250), nullable=False)
 #    vysotnadrm = db.Column(db.String(250), nullable=False)
 #    ploshch = db.Column(db.String(50), nullable=False)
 #    dovz = db.Column(db.String(50), nullable=False)
 #    shyr = db.Column(db.String(50), nullable=False)
-    
-    def __init__(self, name, toponim, type_of_site):
-
-           #  oblast, rajon, punkt, prymitky, kultnal, chron, nadijnist, rozkop, zvit, publicacii, kartograph, coord, tochkart, toppotype, geomorform, vysotnadrm, ploshch, dovz, shyr): 
+#    prymitky = db.Column(db.String(3000))
+#
+    def __init__(self, name, toponim, type_of_site, oblast, rajon): #, punkt, pryvjazka, prymitky, kultnal, localgr, chron, nadijnist, rozkop, zvit, publicacii, kartograph, coord, tochkart, basejn, toppotype, geomorform, vysotnadrm, ploshch, dovz, shyr): 
         self.name = name
         self.toponim = toponim
         self.type_of_site = type_of_site
-#        self.oblast = oblast
-#        self.rajon = rajon
+        self.oblast = oblast
+        self.rajon = rajon
 #        self.punkt = punkt
 #        self.prymitky = prymitky
 #        self.kultnal = kultnal
+#        self.localgr = localgr
 #        self.chron = chron
 #        self.nadijnist = nadijnist
 #        self.rozkop = rozkop
@@ -55,13 +57,14 @@ class Site(db.Model):
 #        self.kartograph = kartograph
 #        self.coord = coord
 #        self.tochkart = tochkart
+#        self.basejn = basejn
 #        self.toppotype = toppotype
 #        self.geomorform = geomorform
 #        self.vysotnadrm = vysotnadrm
-#        self.ploshch = ploshch
+##        self.ploshch = ploshch
 #        self.dovz = dovz
 #        self.shyr= shyr
-
+#
 #    @property #i might have to change this to __init__
 #    def serialize(self):
 #        """Return object data in easily serializeable format"""
@@ -94,30 +97,33 @@ class Site(db.Model):
         return '<Site %r>' % self.name
 
 whooshalchemy.whoosh_index(app, Site)
-@app.route('/search/<query>')
+
+@app.route('/search/<query>', methods=['GET', 'POST'])
 def search(query):
-    #return "hello, world"
-    results = Site.query.whoosh_search(query).all()
-    return render_template('search.html', results=results)
+    if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
+    else:
+        results = Site.query.whoosh_search(query).all()
+        return render_template('search.html', results=results)
 
 @app.route('/', methods=['GET', 'POST'])
 def welcomePage():
     if request.method == 'POST':
         if request.form["name_of_site"]:
             results = request.form["name_of_site"]
-            #results = BlogPost.query.whoosh_search('%s' % search).one()
-            return redirect(url_for('search', query=results)) #, query = search))
-
-            #sSite = session.query(Site).filter_by(name=sname)
-#            site_id=sSite.id 
-#            return redirect(url_for('sitePage', site_id=site_id)) 
-#            print site_id
+            return redirect(url_for('search', query=results)) 
     else:
         return render_template('welcome.html')
 
 @app.route('/new', methods=['GET', 'POST'])
 def newSite():
     if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
+
         TheNewSite = Site(name=request.form['name'], 
                 toponim=request.form['toponim'],
                 type_of_site = request.form['type_of_site'])
@@ -154,14 +160,19 @@ def sitePage(site_id):
         onesite = db.session.query(Site).filter_by(id=site_id).one()
 
         return render_template('site.html', site=onesite) 
-        if request.method == 'POST':
-            return None
+    if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
 
 @app.route('/<int:site_id>/edit/', methods = ['GET', 'POST'])
 @app.route('/<int:site_id>/edit/base', methods = ['GET', 'POST'])
 def siteEdit(site_id):
     siteToEdit = db.session.query(Site).filter_by(id=site_id).one()
     if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
         if request.form['name']:
             siteToEdit.name = request.form['name']
         if request.form['toponim']:
@@ -218,16 +229,24 @@ def siteEdit(site_id):
 def siteDelete(site_id):
     siteToDelete = db.session.query(Site).filter_by(id=site_id).one()
     if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
         db.session.delete(siteToDelete)
         db.session.commit
         return redirect(url_for('welcomePage'))
     else:
         return render_template('delete.html', site=siteToDelete)
 
-@app.route('/all/')
+@app.route('/all/', methods=['GET', 'POST'])
 def allSites():
-    sites = db.session.query(Site).all()
-    return render_template('all.html', sites=sites)
+    if request.method == 'POST':
+        if request.form["name_of_site"]:
+            results = request.form["name_of_site"]
+            return redirect(url_for('search', query=results)) 
+    else:
+        sites = db.session.query(Site).all()
+        return render_template('all.html', sites=sites)
 
 
 if __name__ == '__main__':
